refactor: consolidate W8A8 attention integration
Fold the W8A8 host dispatcher into the operator module and move accuracy and performance coverage into the existing FlashAttention test files.
