--- src/flag_gems/ops/flash_api_w8a8.py
@@ -1,1534 +1,0 @@
-# Copyright 2026 FlagOS Contributors
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import logging
-import math
-
-import torch
-import triton
-
-import flag_gems
-from flag_gems import runtime
-from flag_gems.ops.attention_w8a8_fp8 import (
-    block_m_splitkv_heuristic,
-    block_n_splitkv_heuristic,
-    flash_fwd_kernel,
-    flash_fwd_splitkv_combine_kernel,
-    flash_fwd_splitkv_kernel,
-    flash_varlen_fwd_kernel,
-)
-from flag_gems.runtime import torch_device_fn
-from flag_gems.utils.random_utils import philox_backend_seed_offset
-
-logger = logging.getLogger(__name__)
-_debug = False
-
-
-def CHECK_DEVICE(x):
-    assert x.device.type == flag_gems.device
-
-
-# W8A8 FA3 change: Q/K/V are all FP8 inputs now, while O stays fp16/bf16 like
-# FlashAttention-3 forward.  The fp32 descale tensors below carry the per-block
-# quantization scales used by the QK and PV tensor-core matmuls.
-_FP8_DTYPES = tuple(
-    dtype
-    for dtype in (
-        getattr(torch, "float8_e4m3fn", None),
-        getattr(torch, "float8_e5m2", None),
-        getattr(torch, "float8_e4m3fnuz", None),
-        getattr(torch, "float8_e5m2fnuz", None),
-    )
-    if dtype is not None
-)
-_HIGH_PRECISION_DTYPES = (torch.float16, torch.bfloat16)
-
-
-def _empty_descale(device):
-    # W8A8 FA3 change: use a real empty fp32 tensor instead of None so every
-    # Triton kernel gets the same pointer-bearing argument list.
-    return torch.empty((0,), dtype=torch.float32, device=device)
-
-
-def _default_dense_descale(batch_size, num_heads, nblocks, device):
-    # W8A8 FA3 change: missing descale means the caller already baked the scale
-    # into the FP8 numeric values; use per-block 1.0 scales for compatibility.
-    return torch.ones(
-        (batch_size, num_heads, nblocks), dtype=torch.float32, device=device
-    )
-
-
-def _normalize_dense_descale(descale, batch_size, num_heads, nblocks, device, name):
-    # W8A8 FA3 change: accept FA3-style scalar/per-head/per-block scales and
-    # normalize them to [B, H, nblocks] so the kernels can index one block scale.
-    if descale is None:
-        return _default_dense_descale(batch_size, num_heads, nblocks, device)
-    assert descale.device == device, f"{name} must be on the same device as q"
-    descale = descale.to(torch.float32)
-    if descale.ndim == 0:
-        descale = descale.reshape(1, 1, 1).expand(batch_size, num_heads, nblocks)
-    elif descale.ndim == 1:
-        if descale.numel() == 1:
-            descale = descale.reshape(1, 1, 1).expand(batch_size, num_heads, nblocks)
-        else:
-            assert descale.numel() == num_heads, f"{name} 1D scale must have H elements"
-            descale = descale.reshape(1, num_heads, 1).expand(
-                batch_size, num_heads, nblocks
-            )
-    elif descale.ndim == 2:
-        assert descale.shape == (
-            batch_size,
-            num_heads,
-        ), f"{name} 2D scale must be [B, H]"
-        descale = descale[:, :, None].expand(batch_size, num_heads, nblocks)
-    else:
-        assert descale.shape == (
-            batch_size,
-            num_heads,
-            nblocks,
-        ), f"{name} must be [B, H, nblocks]"
-    return descale.contiguous()
-
-
-class fwd_params:
-    __slots__ = (
-        # pointers and strides
-        "q_ptr",
-        "k_ptr",
-        "v_ptr",
-        "o_ptr",
-        "p_ptr",
-        "softmax_lse_ptr",
-        "q_row_stride",
-        "k_row_stride",
-        "v_row_stride",
-        "q_head_stride",
-        "k_head_stride",
-        "v_head_stride",
-        "o_row_stride",
-        "o_head_stride",
-        "q_batch_stride",
-        "k_batch_stride",
-        "v_batch_stride",
-        "o_batch_stride",
-        "is_cu_seqlens_q",
-        "cu_seqlens_q_ptr",
-        "is_cu_seqlens_k",
-        "cu_seqlens_k_ptr",
-        "is_seqused_k",
-        "seqused_k_ptr",
-        # sizes
-        "b",
-        "bk",
-        "h",
-        "hk",
-        "h_hk_ratio",
-        "seqlen_q",
-        "seqlen_k",
-        "seqlen_q_rounded",
-        "seqlen_k_rounded",
-        "d",
-        "d_rounded",
-        # scaling factors
-        "is_softcap",
-        "softcap",
-        "scale_softmax",
-        "scale_softmax_log2",
-        "q_descale_ptr",
-        "k_descale_ptr",
-        "v_descale_ptr",
-        "q_descale_batch_stride",
-        "q_descale_head_stride",
-        "q_descale_block_stride",
-        "k_descale_batch_stride",
-        "k_descale_head_stride",
-        "k_descale_block_stride",
-        "v_descale_batch_stride",
-        "v_descale_head_stride",
-        "v_descale_block_stride",
-        "use_fa3_fp8_scales",
-        "fp8_p_max",
-        # dropout
-        "is_dropout",
-        "p_dropout",
-        "rp_dropout",
-        "p_dropout_in_uint8_t",
-        "philox_args",
-        "return_softmax",
-        # masking
-        "is_causal",
-        "is_local",
-        "window_size_left",
-        "window_size_right",
-        "seqlenq_ngroups_swapped",
-        "is_paged",
-        # alibi
-        "is_alibi",
-        "alibi_slopes_ptr",
-        "alibi_slopes_batch_stride",
-        # block table
-        "total_q",
-        "page_table_ptr",
-        "page_table_batch_stride",
-        "block_size",
-    )
-
-    def __init__(
-        self,
-        q_ptr,
-        k_ptr,
-        v_ptr,
-        o_ptr,
-        p_ptr,
-        softmax_lse_ptr,
-        q_row_stride,
-        k_row_stride,
-        v_row_stride,
-        q_head_stride,
-        k_head_stride,
-        v_head_stride,
-        o_row_stride,
-        o_head_stride,
-        q_batch_stride,
-        k_batch_stride,
-        v_batch_stride,
-        o_batch_stride,
-        is_cu_seqlens_q,
-        cu_seqlens_q_ptr,
-        is_cu_seqlens_k,
-        cu_seqlens_k_ptr,
-        is_seqused_k,
-        seqused_k_ptr,
-        # sizes
-        b,
-        bk,
-        h,
-        hk,
-        h_hk_ratio,
-        seqlen_q,
-        seqlen_k,
-        seqlen_q_rounded,
-        seqlen_k_rounded,
-        d,
-        d_rounded,
-        # scaling factors
-        is_softcap,
-        softcap,
-        scale_softmax,
-        scale_softmax_log2,
-        q_descale_ptr,
-        k_descale_ptr,
-        v_descale_ptr,
-        q_descale_batch_stride,
-        q_descale_head_stride,
-        q_descale_block_stride,
-        k_descale_batch_stride,
-        k_descale_head_stride,
-        k_descale_block_stride,
-        v_descale_batch_stride,
-        v_descale_head_stride,
-        v_descale_block_stride,
-        use_fa3_fp8_scales,
-        fp8_p_max,
-        # dropout
-        is_dropout,
-        p_dropout,
-        rp_dropout,
-        p_dropout_in_uint8_t,
-        philox_args,
-        return_softmax,
-        # masking
-        is_causal,
-        is_local,
-        window_size_left,
-        window_size_right,
-        seqlenq_ngroups_swapped,
-        is_paged,
-        # alibi
-        is_alibi,
-        alibi_slopes_ptr,
-        alibi_slopes_batch_stride,
-        # block table
-        total_q,
-        page_table_ptr,
-        page_table_batch_stride,
-        block_size,
-    ):
-        self.q_ptr = q_ptr
-        self.k_ptr = k_ptr
-        self.v_ptr = v_ptr
-        self.o_ptr = o_ptr
-        self.p_ptr = p_ptr
-        self.softmax_lse_ptr = softmax_lse_ptr
-        self.q_row_stride = q_row_stride
-        self.k_row_stride = k_row_stride
-        self.v_row_stride = v_row_stride
-        self.q_head_stride = q_head_stride
-        self.k_head_stride = k_head_stride
-        self.v_head_stride = v_head_stride
-        self.o_row_stride = o_row_stride
-        self.o_head_stride = o_head_stride
-        self.q_batch_stride = q_batch_stride
-        self.k_batch_stride = k_batch_stride
-        self.v_batch_stride = v_batch_stride
-        self.o_batch_stride = o_batch_stride
-        self.is_cu_seqlens_q = is_cu_seqlens_q
-        self.cu_seqlens_q_ptr = cu_seqlens_q_ptr
-        self.is_cu_seqlens_k = is_cu_seqlens_k
-        self.cu_seqlens_k_ptr = cu_seqlens_k_ptr
-        self.is_seqused_k = is_seqused_k
-        self.seqused_k_ptr = seqused_k_ptr
-        # sizes
-        self.b = b
-        self.bk = bk
-        self.h = h
-        self.hk = hk
-        self.h_hk_ratio = h_hk_ratio
-        self.seqlen_q = seqlen_q
-        self.seqlen_k = seqlen_k
-        self.seqlen_q_rounded = seqlen_q_rounded
-        self.seqlen_k_rounded = seqlen_k_rounded
-        self.d = d
-        self.d_rounded = d_rounded
-        # scaling factors
-        self.is_softcap = is_softcap
-        self.softcap = softcap
-        self.scale_softmax = scale_softmax
-        self.scale_softmax_log2 = scale_softmax_log2
-        self.q_descale_ptr = q_descale_ptr
-        self.k_descale_ptr = k_descale_ptr
-        self.v_descale_ptr = v_descale_ptr
-        self.q_descale_batch_stride = q_descale_batch_stride
-        self.q_descale_head_stride = q_descale_head_stride
-        self.q_descale_block_stride = q_descale_block_stride
-        self.k_descale_batch_stride = k_descale_batch_stride
-        self.k_descale_head_stride = k_descale_head_stride
-        self.k_descale_block_stride = k_descale_block_stride
-        self.v_descale_batch_stride = v_descale_batch_stride
-        self.v_descale_head_stride = v_descale_head_stride
-        self.v_descale_block_stride = v_descale_block_stride
-        self.use_fa3_fp8_scales = use_fa3_fp8_scales
-        self.fp8_p_max = fp8_p_max
-        # dropout
-        self.is_dropout = is_dropout
-        self.p_dropout = p_dropout
-        self.rp_dropout = rp_dropout
-        self.p_dropout_in_uint8_t = p_dropout_in_uint8_t
-        self.philox_args = philox_args
-        self.return_softmax = return_softmax
-        # masking
-        self.is_causal = is_causal
-        self.is_local = is_local
-        self.window_size_left = window_size_left
-        self.window_size_right = window_size_right
-        self.seqlenq_ngroups_swapped = seqlenq_ngroups_swapped
-        self.is_paged = is_paged
-        # alibi
-        self.is_alibi = is_alibi
-        self.alibi_slopes_ptr = alibi_slopes_ptr
-        self.alibi_slopes_batch_stride = alibi_slopes_batch_stride
-        # block table
-        self.total_q = total_q
-        self.page_table_ptr = page_table_ptr
-        self.page_table_batch_stride = page_table_batch_stride
-        self.block_size = block_size
-
-    def args(self):
-        return tuple(getattr(self, k) for k in self.__slots__)
-
-
-def mha_varlan_fwd(
-    q,
-    k,
-    v,
-    out,
-    cu_seqlens_q,
-    cu_seqlens_k,
-    seqused_k,
-    leftpad_k,
-    page_table,
-    alibi_slopes,
-    max_seqlen_q,
-    max_seqlen_k,
-    p_dropout,
-    softmax_scale,
-    zero_tensors,
-    is_causal,
-    window_size_left,
-    window_size_right,
-    softcap,
-    return_softmax,
-    gen,
-    q_descale=None,
-    k_descale=None,
-    v_descale=None,
-    fp8_p_max=448.0,
-):
-    CHECK_DEVICE(q), CHECK_DEVICE(k), CHECK_DEVICE(v)
-    q_device = q.device
-    q_dtype = q.dtype
-    v_dtype = v.dtype
-    # W8A8 FA3 change: varlen forward also consumes FP8 Q/K/V and uses explicit
-    # descale tensors, matching the dense kernel contract.
-    assert q_dtype in _FP8_DTYPES, "W8A8 FA3 FlashAttention expects q to be fp8"
-    assert (
-        k.dtype == q_dtype
-    ), "W8A8 FA3 FlashAttention expects q and k to use the same fp8 dtype"
-    assert v_dtype == q_dtype, "W8A8 FA3 FlashAttention expects v to be fp8"
-    assert q.stride(-1) == 1, "Input tensor must have contiguous last dimension"
-    assert k.stride(-1) == 1, "Input tensor must have contiguous last dimension"
-    assert v.stride(-1) == 1, "Input tensor must have contiguous last dimension"
-
-    assert cu_seqlens_q.dtype == torch.int32
-    assert cu_seqlens_q.is_contiguous()
-
-    assert cu_seqlens_k.dtype == torch.int32
-    assert cu_seqlens_k.is_contiguous()
-
-    is_paged = page_table is not None
-    if not is_paged:
-        page_table = torch.empty((0, 0), device=q_device, dtype=torch.int32)
-
-    # q shape: [total_q_tokens, num_heads, head_size]
-    # k shape:
-    #   paged_kv: [num_pages, block_size, num_heads_k, head_size]
-    # batch_size, number of sentences
-    total_q, num_heads, head_size = q.size()
-    num_heads_k = k.size(2) if is_paged else k.size(1)
-    batch_size = cu_seqlens_q.numel() - 1
-    block_size = k.size(1) if is_paged else 1
-    num_pages = k.size(0) if is_paged else 0
-    k_batch_size = num_pages
-    # max_num_pages_per_seq = page_table.size(1)
-    page_table_batch_stride = page_table.stride(0)
-    k_batch_stride = k.stride(0)
-    v_batch_stride = v.stride(0)
-
-    assert k.size() == v.size()
-    assert cu_seqlens_q.size() == (batch_size + 1,)
-    assert cu_seqlens_k.size() == (batch_size + 1,)
-
-    # Check output shape
-    if out is not None:
-        assert out.stride(-1) == 1
-        # W8A8 FA3 change: output stays high precision although V is FP8.
-        assert out.dtype in _HIGH_PRECISION_DTYPES
-        assert out.size() == (total_q, num_heads, head_size)
-
-    if seqused_k is not None:
-        assert seqused_k.is_contiguous()
-        assert seqused_k.size() == (batch_size,)
-
-    if max_seqlen_q == 1 and alibi_slopes is None:
-        is_causal = False
-
-    if is_causal:
-        window_size_right = 0
-
-    # check disable swa
-    if window_size_left >= max_seqlen_k:
-        window_size_left = -1
-    if window_size_right >= max_seqlen_k:
-        window_size_right = -1
-
-    is_local = window_size_left >= 0
-
-    # Optimize all single-query sequences by swapping the query-group and sequence dimensions
-    seqlenq_ngroups_swapped = (
-        max_seqlen_q == 1
-        and alibi_slopes is None
-        and num_heads > num_heads_k
-        and window_size_left < 0
-        and window_size_right < 0
-        and p_dropout == 0
-    )
-    q_groups = num_heads // num_heads_k
-    if seqlenq_ngroups_swapped:
-        logger.debug("Swapping query groups and sequence dimensions")
-        q = (
-            q.reshape((batch_size, num_heads_k, q_groups, head_size))
-            .transpose(1, 2)
-            .reshape(batch_size * q_groups, num_heads_k, head_size)
-        )
-        max_seqlen_q = q_groups
-        num_heads = num_heads_k
-        cu_seqlens_q = None
-        q_batch_stride = q.stride(0) * max_seqlen_q
-        k_batch_stride = k.stride(0)
-        v_batch_stride = v.stride(0)
-        # o_batch_stride = out.stride(0) * max_seqlen_q
-    else:
-        q_batch_stride = 0
-        k_batch_stride = 0
-        v_batch_stride = 0
-        o_batch_stride = 0
-
-    total_q = q.size(0)
-
-    assert leftpad_k is None, "leftpad_k is not supported."
-    assert (
-        head_size <= 256
-    ), "FlashAttention forward only supports head dimension at most 256"
-    assert (
-        head_size % 8 == 0
-    ), "head_size must be a multiple of 8, this is ensured by padding!"
-    assert (
-        num_heads % num_heads_k == 0
-    ), "Number of heads in key/value must divide number of heads in query"
-
-    assert q.shape == (total_q, num_heads, head_size)
-    if is_paged:
-        assert k.shape == (num_pages, block_size, num_heads_k, head_size)
-        assert v.shape == (num_pages, block_size, num_heads_k, head_size)
-    assert k.stride() == v.stride()
-
-    if softcap > 0.0:
-        assert p_dropout == 0, "dropout is not supported if softcap is used."
-
-    round_multiple = lambda x, m: (x + m - 1) // m * m
-    head_size_rounded = round_multiple(head_size, 32) if head_size <= 192 else 256
-    seqlen_q_rounded = round_multiple(max_seqlen_q, 128)
-    seqlen_k_rounded = round_multiple(max_seqlen_k, 32)
-
-    M_LOG2E = 1.4426950408889634074
-    if softcap > 0.0:
-        is_softcap = True
-        adjusted_scale_softmax = softcap
-        adjusted_softcap = softmax_scale / softcap
-        adjusted_scale_softmax_log2e = softcap * M_LOG2E
-    else:
-        is_softcap = False
-        adjusted_softcap = 0.0
-        adjusted_scale_softmax = softmax_scale
-        adjusted_scale_softmax_log2e = softmax_scale * M_LOG2E
-
-    # Set alibi params
-    if alibi_slopes is not None:
-        assert alibi_slopes.device == q_device
-        assert alibi_slopes.dtype in (torch.float,)
-        assert alibi_slopes.stride(-1) == 1
-        assert alibi_slopes.shape == (num_heads,) or alibi_slopes.shape == (
-            batch_size,
-            num_heads,
-        )
-        alibi_slopes_batch_stride = (
-            alibi_slopes.stride(0) if alibi_slopes.ndim == 2 else 0
-        )
-        is_alibi = True
-    else:
-        alibi_slopes_batch_stride = 0
-        is_alibi = False
-
-    # Prepare params to kernel
-    with torch_device_fn.device(q_device):
-        if out is not None:
-            out_ = out
-            if seqlenq_ngroups_swapped:
-                out = torch.empty_like(q, dtype=torch.bfloat16)
-        else:
-            out_ = None
-            out = torch.empty_like(q, dtype=torch.bfloat16)
-
-        if seqlenq_ngroups_swapped:
-            o_batch_stride = out.stride(0) * max_seqlen_q
-
-        lse = torch.empty((num_heads, total_q), dtype=torch.float, device=q_device)
-
-        if p_dropout > 0:
-            is_dropout = True
-            increment = batch_size * num_heads * 32
-            philox_seed, philox_offset = philox_backend_seed_offset(increment)
-            philox_args = torch.tensor(
-                [philox_seed, philox_offset], dtype=torch.int64, device=q_device
-            )
-        else:
-            is_dropout = False
-            philox_args = torch.empty((2,), dtype=torch.int64, device=q_device)
-
-        p_dropout = 1 - p_dropout
-        p_dropout_in_uint8_t = math.floor(p_dropout * 255.0)
-        rp_dropout = 1.0 / p_dropout
-
-        if return_softmax:
-            assert is_dropout, "Only supported with non-zero dropout."
-            p = torch.empty(
-                (batch_size, num_heads, seqlen_q_rounded, seqlen_k_rounded),
-                device=q_device,
-            )
-        else:
-            p = torch.empty((), device=q_device)
-
-        if zero_tensors:
-            out.zero_()
-            lse.fill_(float("-inf"))
-
-        q_nblocks = triton.cdiv(max_seqlen_q, 128)
-        k_nblocks = triton.cdiv(max_seqlen_k, 128)
-        # W8A8 FA3 change: varlen kernels use [B, H, block] descales indexed by
-        # request id and local block id.  Paged KV uses the same logical blocks.
-        q_descale = _normalize_dense_descale(
-            q_descale, batch_size, num_heads, q_nblocks, q_device, "q_descale"
-        )
-        k_descale = _normalize_dense_descale(
-            k_descale, batch_size, num_heads_k, k_nblocks, q_device, "k_descale"
-        )
-        v_descale = _normalize_dense_descale(
-            v_descale, batch_size, num_heads_k, k_nblocks, q_device, "v_descale"
-        )
-
-        params = fwd_params(
-            q,  # q_ptr,
-            k,  # k_ptr,
-            v,  # v_ptr,
-            out,  # o_ptr,
-            p,  # p_ptr,
-            lse,  # softmax_lse_ptr,
-            q.stride(-3),  # q_row_stride,
-            k.stride(-3),  # k_row_stride,
-            v.stride(-3),  # v_row_stride,
-            q.stride(-2),  # q_head_stride,
-            k.stride(-2),  # k_head_stride,
-            v.stride(-2),  # v_head_stride,
-            out.stride(-3),  # o_row_stride,
-            out.stride(-2),  # o_head_stride,
-            q_batch_stride,  # q_batch_stride,
-            k_batch_stride,  # k_batch_stride,
-            v_batch_stride,  # v_batch_stride,
-            o_batch_stride,  # o_batch_stride,
-            cu_seqlens_q is not None,  # is_cu_seqlens_q,
-            cu_seqlens_q,  # cu_seqlens_q_ptr,
-            seqused_k is None,  # is_cu_seqlens_k,
-            cu_seqlens_k,  # cu_seqlens_k_ptr,
-            seqused_k is not None,  # is_seqused_k,
-            seqused_k,  # seqused_k_ptr,
-            # sizes
-            batch_size,  # b,
-            k_batch_size,  # bk,
-            num_heads,  # h,
-            num_heads_k,  # hk,
-            num_heads // num_heads_k,  # h_hk_ratio,
-            max_seqlen_q,  # seqlen_q,
-            max_seqlen_k,  # seqlen_k,
-            seqlen_q_rounded,  # seqlen_q_rounded,
-            seqlen_k_rounded,  # seqlen_k_rounded,
-            head_size,  # d,
-            head_size_rounded,  # d_rounded,
-            # scaling factors
-            is_softcap,
-            adjusted_softcap,  # softcap,
-            adjusted_scale_softmax,  # scale_softmax,
-            adjusted_scale_softmax_log2e,  # scale_softmax_log2,
-            q_descale,  # q_descale_ptr,
-            k_descale,  # k_descale_ptr,
-            v_descale,  # v_descale_ptr,
-            q_descale.stride(0),  # q_descale_batch_stride,
-            q_descale.stride(1),  # q_descale_head_stride,
-            q_descale.stride(2),  # q_descale_block_stride,
-            k_descale.stride(0),  # k_descale_batch_stride,
-            k_descale.stride(1),  # k_descale_head_stride,
-            k_descale.stride(2),  # k_descale_block_stride,
-            v_descale.stride(0),  # v_descale_batch_stride,
-            v_descale.stride(1),  # v_descale_head_stride,
-            v_descale.stride(2),  # v_descale_block_stride,
-            True,  # use_fa3_fp8_scales,
-            fp8_p_max,  # fp8_p_max,
-            # dropout
-            is_dropout,
-            p_dropout,
-            rp_dropout,
-            p_dropout_in_uint8_t,
-            philox_args,
-            return_softmax,
-            # causal and swa
-            is_causal,  # is_causal,
-            is_local,  # is_local,
-            window_size_left,  # window_size_left,
-            window_size_right,  # window_size_right,
-            seqlenq_ngroups_swapped,  # seqlenq_ngroups_swapped,
-            is_paged,
-            # alibi
-            is_alibi,  #
-            alibi_slopes,  # alibi_slopes_ptr,
-            alibi_slopes_batch_stride,  # alibi_slopes_batch_stride,
-            # block table params
-            total_q,  # total_q,
-            page_table,  # page_table_ptr,
-            page_table_batch_stride,  # page_table_batch_stride,
-            block_size,  # block_size,
-        )
-
-        if flag_gems.vendor_name == "iluvatar":
-            params.k_ptr = k.view(k.shape[0], k.shape[1], -1)
-            params.v_ptr = v.view(v.shape[0], v.shape[1], -1)
-        logger.debug("kernel: flash_varlen_fwd")
-        # 改动：D128 非 paged varlen 也启用 2-way split-D；paged cache loader 暂保持原 D 维完整加载。
-        use_varlen_split_d = head_size == 128 and not is_paged
-        grid = lambda args: (
-            triton.cdiv(max_seqlen_q, args["BLOCK_M"]),
-            batch_size,
-            num_heads * (2 if use_varlen_split_d else 1),
-        )
-        kernel = flash_varlen_fwd_kernel[grid]
-        args = tuple(getattr(params, k) for k in params.__slots__)
-
-        # We assess which phase the requests are likely to be in and set the config accordingly.
-        total_rows = total_q * num_heads
-        num_sms = torch_device_fn.get_device_properties(
-            flag_gems.device
-        ).multi_processor_count
-        avg_rows_per_sm = total_rows / num_sms
-        avg_rows_per_batch = total_q / batch_size
-        avg_rows_per_cta = min(avg_rows_per_batch, avg_rows_per_sm)
-        # Heuristic: if avg_rows_per_sm >= 128, we are likely in prefill phase.
-        # This is a rough heuristic and may not be accurate for all scenarios.
-        if avg_rows_per_cta > 64:
-            varlen_fwd_config_str = "mha_block_128"
-        elif avg_rows_per_cta > 32:
-            varlen_fwd_config_str = "mha_block_64"
-        elif avg_rows_per_cta > 16:
-            varlen_fwd_config_str = "mha_block_32"
-        else:
-            varlen_fwd_config_str = "mha_block_16"
-        if flag_gems.vendor_name == "mthreads":
-            varlen_fwd_config_str = "mha_block_32"
-
-        cfg = runtime.get_heuristic_config(varlen_fwd_config_str)
-        cfg_params = {
-            "BLOCK_M": cfg["BLOCK_M"](args),
-            "BLOCK_N": cfg["BLOCK_N"](args),
-            "BLOCK_K": triton.next_power_of_2(head_size),
-            # 改动：QK 仍使用完整 BLOCK_K；PV/output 在 D128 split-D 时使用 BLOCK_D=64。
-            "BLOCK_D": 64 if use_varlen_split_d else triton.next_power_of_2(head_size),
-            "SPLIT_D": use_varlen_split_d,
-            "num_warps": cfg["num_warps"](args),
-            "num_stages": 1 if not is_paged else cfg["num_stages"](args),
-        }
-
-        logger.debug("Running flash_varlen_fwd_kernel with config: %s", cfg_params)
-        kernel(*args, **cfg_params)
-
-        if seqlenq_ngroups_swapped:
-            out = out.reshape(
-                batch_size, max_seqlen_q, num_heads_k, head_size
-            ).transpose(1, 2)
-            if out_ is not None:
-                out_.view(batch_size, num_heads_k, max_seqlen_q, head_size).copy_(out)
-                out = out_
-            else:
-                out = out.reshape(batch_size, num_heads_k * max_seqlen_q, head_size)
-            lse = lse.reshape(num_heads_k, batch_size, max_seqlen_q)
-            lse = lse.reshape(num_heads_k * max_seqlen_q, batch_size)
-
-        unused = torch.empty((), dtype=torch.int64, device=q_device)
-    return out, q, k, v, lse, philox_args, unused, p
-
-
-def mha_varlan_fwd_opt(
-    q,
-    k,
-    v,
-    out,
-    lse,
-    cu_seqlens_q,
-    cu_seqlens_k,
-    seqused_k,
-    leftpad_k,
-    page_table,
-    alibi_slopes,
-    max_seqlen_q,
-    max_seqlen_k,
-    p_dropout,
-    softmax_scale,
-    zero_tensors,
-    is_causal,
-    window_size_left,
-    window_size_right,
-    softcap,
-    return_softmax,
-    gen,
-    q_descale=None,
-    k_descale=None,
-    v_descale=None,
-    fp8_p_max=448.0,
-):
-    CHECK_DEVICE(q), CHECK_DEVICE(k), CHECK_DEVICE(v)
-    q_device = q.device
-    q_dtype = q.dtype
-    v_dtype = v.dtype
-    # W8A8 FA3 change: optimized varlen path uses FP8 Q/K/V and per-block
-    # descales just like FA3 FP8 forward.
-    assert q_dtype in _FP8_DTYPES, "W8A8 FA3 FlashAttention expects q to be fp8"
-    assert (
-        k.dtype == q_dtype
-    ), "W8A8 FA3 FlashAttention expects q and k to use the same fp8 dtype"
-    assert v_dtype == q_dtype, "W8A8 FA3 FlashAttention expects v to be fp8"
-    assert q.stride(-1) == 1, "Input tensor must have contiguous last dimension"
-    assert k.stride(-1) == 1, "Input tensor must have contiguous last dimension"
-    assert v.stride(-1) == 1, "Input tensor must have contiguous last dimension"
-
-    assert cu_seqlens_q.dtype == torch.int32
-    assert cu_seqlens_q.is_contiguous()
-
-    assert cu_seqlens_k.dtype == torch.int32
-    assert cu_seqlens_k.is_contiguous()
-
-    is_paged = page_table is not None
-    if not is_paged:
-        page_table = torch.empty((0, 0), device=q_device, dtype=torch.int32)
-
-    # q shape: [total_q_tokens, num_heads, head_size]
-    # k shape:
-    #   paged_kv: [num_pages, block_size, num_heads_k, head_size]
-    # batch_size, number of sentences
-    total_q, num_heads, head_size = q.size()
-    num_heads_k = k.size(2) if is_paged else k.size(1)
-    batch_size = cu_seqlens_q.numel() - 1
-    block_size = k.size(1) if is_paged else 1
-    num_pages = k.size(0) if is_paged else 0
-    k_batch_size = num_pages
-    # max_num_pages_per_seq = page_table.size(1)
-    page_table_batch_stride = page_table.stride(0)
-    k_batch_stride = k.stride(0)
-    v_batch_stride = v.stride(0)
-
-    assert k.size() == v.size()
-    assert cu_seqlens_q.size() == (batch_size + 1,)
-    assert cu_seqlens_k.size() == (batch_size + 1,)
-
-    # Check output shape
-    if out is not None:
-        assert out.stride(-1) == 1
-        # W8A8 FA3 change: optimized varlen output remains high precision.
-        assert out.dtype in _HIGH_PRECISION_DTYPES
-        assert out.size() == (total_q, num_heads, head_size)
-
-    if seqused_k is not None:
-        assert seqused_k.is_contiguous()
-        assert seqused_k.size() == (batch_size,)
-
-    if max_seqlen_q == 1 and alibi_slopes is None:
-        is_causal = False
-
-    if is_causal:
-        window_size_right = 0
-
-    # check disable swa
-    if window_size_left >= max_seqlen_k:
-        window_size_left = -1
-    if window_size_right >= max_seqlen_k:
-        window_size_right = -1
-
-    is_local = window_size_left >= 0
-
-    # Optimize all single-query sequences by swapping the query-group and sequence dimensions
-    seqlenq_ngroups_swapped = (
-        max_seqlen_q == 1
-        and alibi_slopes is None
-        and num_heads > num_heads_k
-        and window_size_left < 0
-        and window_size_right < 0
-        and p_dropout == 0
-    )
-    q_groups = num_heads // num_heads_k
-    if seqlenq_ngroups_swapped:
-        logger.debug("Swapping query groups and sequence dimensions")
-        q = (
-            q.reshape((batch_size, num_heads_k, q_groups, head_size))
-            .transpose(1, 2)
-            .reshape(batch_size * q_groups, num_heads_k, head_size)
-        )
-        max_seqlen_q = q_groups
-        num_heads = num_heads_k
-        cu_seqlens_q = None
-        q_batch_stride = q.stride(0) * max_seqlen_q
-        k_batch_stride = k.stride(0)
-        v_batch_stride = v.stride(0)
-        # o_batch_stride = out.stride(0) * max_seqlen_q
-    else:
-        q_batch_stride = 0
-        k_batch_stride = 0
-        v_batch_stride = 0
-        o_batch_stride = 0
-
-    total_q = q.size(0)
-
-    assert leftpad_k is None, "leftpad_k is not supported."
-    assert (
-        head_size <= 256
-    ), "FlashAttention forward only supports head dimension at most 256"
-    assert (
-        head_size % 8 == 0
-    ), "head_size must be a multiple of 8, this is ensured by padding!"
-    assert (
-        num_heads % num_heads_k == 0
-    ), "Number of heads in key/value must divide number of heads in query"
-
-    assert q.shape == (total_q, num_heads, head_size)
-    if is_paged:
-        assert k.shape == (num_pages, block_size, num_heads_k, head_size)
-        assert v.shape == (num_pages, block_size, num_heads_k, head_size)
-    assert k.stride() == v.stride()
-
-    if softcap > 0.0:
-        assert p_dropout == 0, "dropout is not supported if softcap is used."
-
-    round_multiple = lambda x, m: (x + m - 1) // m * m
-    head_size_rounded = round_multiple(head_size, 32) if head_size <= 192 else 256
-    seqlen_q_rounded = round_multiple(max_seqlen_q, 128)
-    seqlen_k_rounded = round_multiple(max_seqlen_k, 32)
-
-    M_LOG2E = 1.4426950408889634074
-    if softcap > 0.0:
-        is_softcap = True
-        adjusted_scale_softmax = softcap
-        adjusted_softcap = softmax_scale / softcap
-        adjusted_scale_softmax_log2e = softcap * M_LOG2E
-    else:
-        is_softcap = False
-        adjusted_softcap = 0.0
-        adjusted_scale_softmax = softmax_scale
-        adjusted_scale_softmax_log2e = softmax_scale * M_LOG2E
-
-    # Set alibi params
-    if alibi_slopes is not None:
-        assert alibi_slopes.device == q_device
-        assert alibi_slopes.dtype in (torch.float,)
-        assert alibi_slopes.stride(-1) == 1
-        assert alibi_slopes.shape == (num_heads,) or alibi_slopes.shape == (
-            batch_size,
-            num_heads,
-        )
-        alibi_slopes_batch_stride = (
-            alibi_slopes.stride(0) if alibi_slopes.ndim == 2 else 0
-        )
-        is_alibi = True
-    else:
-        alibi_slopes_batch_stride = 0
-        is_alibi = False
-
-    # Prepare params to kernel
-    with torch_device_fn.device(q_device):
-        if out is not None:
-            out_ = out
-            if seqlenq_ngroups_swapped:
-                # W8A8 FA3 change: temporary output is high precision; V is FP8.
-                out = torch.empty_like(q, dtype=torch.bfloat16)
-        else:
-            out_ = None
-            # W8A8 FA3 change: optimized varlen creates high-precision output
-            # even when the caller passes FP8 Q/K/V.
-            out = torch.empty_like(q, dtype=torch.bfloat16)
-
-        if seqlenq_ngroups_swapped:
-            o_batch_stride = out.stride(0) * max_seqlen_q
-
-        if lse is None:
-            lse = torch.empty((num_heads, total_q), dtype=torch.float, device=q_device)
-
-        if p_dropout > 0:
-            is_dropout = True
-            increment = batch_size * num_heads * 32
-            philox_seed, philox_offset = philox_backend_seed_offset(increment)
-            philox_args = torch.tensor(
-                [philox_seed, philox_offset], dtype=torch.int64, device=q_device
-            )
-        else:
-            is_dropout = False
-            # philox_args = torch.empty((2,), dtype=torch.int64, device=q_device)
-            philox_args = None
-
-        p_dropout = 1 - p_dropout
-        p_dropout_in_uint8_t = math.floor(p_dropout * 255.0)
-        rp_dropout = 1.0 / p_dropout
-
-        if return_softmax:
-            assert is_dropout, "Only supported with non-zero dropout."
-            p = torch.empty(
-                (batch_size, num_heads, seqlen_q_rounded, seqlen_k_rounded),
-                device=q_device,
-            )
-        else:
-            # p = torch.empty((), device=q_device)
-            p = None
-        if zero_tensors:
-            out.zero_()
-            lse.fill_(float("-inf"))
-
-        q_nblocks = triton.cdiv(max_seqlen_q, 128)
-        k_nblocks = triton.cdiv(max_seqlen_k, 128)
-        # W8A8 FA3 change: normalize optional FA3 descales for the optimized
-        # varlen path before packing the common fwd_params object.
-        q_descale = _normalize_dense_descale(
-            q_descale, batch_size, num_heads, q_nblocks, q_device, "q_descale"
-        )
-        k_descale = _normalize_dense_descale(
-            k_descale, batch_size, num_heads_k, k_nblocks, q_device, "k_descale"
-        )
-        v_descale = _normalize_dense_descale(
-            v_descale, batch_size, num_heads_k, k_nblocks, q_device, "v_descale"
-        )
-
-        params = fwd_params(
-            q,  # q_ptr,
-            k,  # k_ptr,
-            v,  # v_ptr,
-            out,  # o_ptr,
-            p,  # p_ptr,
-            lse,  # softmax_lse_ptr,
-            q.stride(-3),  # q_row_stride,
-            k.stride(-3),  # k_row_stride,
-            v.stride(-3),  # v_row_stride,
-            q.stride(-2),  # q_head_stride,
-            k.stride(-2),  # k_head_stride,
-            v.stride(-2),  # v_head_stride,
-            out.stride(-3),  # o_row_stride,
-            out.stride(-2),  # o_head_stride,
-            q_batch_stride,  # q_batch_stride,
-            k_batch_stride,  # k_batch_stride,
-            v_batch_stride,  # v_batch_stride,
-            o_batch_stride,  # o_batch_stride,
-            cu_seqlens_q is not None,  # is_cu_seqlens_q,
-            cu_seqlens_q,  # cu_seqlens_q_ptr,
-            cu_seqlens_k is not None,  # is_cu_seqlens_k,
-            cu_seqlens_k,  # cu_seqlens_k_ptr,
-            seqused_k is not None,  # is_seqused_k,
-            seqused_k,  # seqused_k_ptr,
-            # sizes
-            batch_size,  # b,
-            k_batch_size,  # bk,
-            num_heads,  # h,
-            num_heads_k,  # hk,
-            num_heads // num_heads_k,  # h_hk_ratio,
-            max_seqlen_q,  # seqlen_q,
-            max_seqlen_k,  # seqlen_k,
-            seqlen_q_rounded,  # seqlen_q_rounded,
-            seqlen_k_rounded,  # seqlen_k_rounded,
-            head_size,  # d,
-            head_size_rounded,  # d_rounded,
-            # scaling factors
-            is_softcap,
-            adjusted_softcap,  # softcap,
-            adjusted_scale_softmax,  # scale_softmax,
-            adjusted_scale_softmax_log2e,  # scale_softmax_log2,
-            q_descale,  # q_descale_ptr,
-            k_descale,  # k_descale_ptr,
-            v_descale,  # v_descale_ptr,
-            q_descale.stride(0),  # q_descale_batch_stride,
-            q_descale.stride(1),  # q_descale_head_stride,
-            q_descale.stride(2),  # q_descale_block_stride,
-            k_descale.stride(0),  # k_descale_batch_stride,
-            k_descale.stride(1),  # k_descale_head_stride,
-            k_descale.stride(2),  # k_descale_block_stride,
-            v_descale.stride(0),  # v_descale_batch_stride,
-            v_descale.stride(1),  # v_descale_head_stride,
-            v_descale.stride(2),  # v_descale_block_stride,
-            True,  # use_fa3_fp8_scales,
-            fp8_p_max,  # fp8_p_max,
-            # dropout
-            is_dropout,
-            p_dropout,
-            rp_dropout,
-            p_dropout_in_uint8_t,
-            philox_args,
-            return_softmax,
-            # causal and swa
-            is_causal,  # is_causal,
-            is_local,  # is_local,
-            window_size_left,  # window_size_left,
-            window_size_right,  # window_size_right,
-            seqlenq_ngroups_swapped,  # seqlenq_ngroups_swapped,
-            is_paged,
-            # alibi
-            is_alibi,  #
-            alibi_slopes,  # alibi_slopes_ptr,
-            alibi_slopes_batch_stride,  # alibi_slopes_batch_stride,
-            # block table params
-            total_q,  # total_q,
-            page_table,  # page_table_ptr,
-            page_table_batch_stride,  # page_table_batch_stride,
-            block_size,  # block_size,
-        )
-
-        if flag_gems.vendor_name == "iluvatar":
-            params.k_ptr = k.view(k.shape[0], k.shape[1], -1)
-            params.v_ptr = v.view(v.shape[0], v.shape[1], -1)
-        logger.debug("kernel: flash_varlen_fwd")
-        # 改动：D128 非 paged varlen 也启用 2-way split-D；paged cache loader 暂保持原 D 维完整加载。
-        use_varlen_split_d = head_size == 128 and not is_paged
-        grid = lambda args: (
-            triton.cdiv(max_seqlen_q, args["BLOCK_M"]),
-            batch_size,
-            num_heads * (2 if use_varlen_split_d else 1),
-        )
-        kernel = flash_varlen_fwd_kernel[grid]
-        args = tuple(getattr(params, k) for k in params.__slots__)
-
-        # We assess which phase the requests are likely to be in and set the config accordingly.
-        total_rows = total_q * num_heads
-        num_sms = torch_device_fn.get_device_properties(
-            flag_gems.device
-        ).multi_processor_count
-        avg_rows_per_sm = total_rows / num_sms
-        avg_rows_per_batch = total_q / batch_size
-        avg_rows_per_cta = min(avg_rows_per_batch, avg_rows_per_sm)
-        # Heuristic: if avg_rows_per_sm >= 128, we are likely in prefill phase.
-        # This is a rough heuristic and may not be accurate for all scenarios.
-        if avg_rows_per_cta > 64:
-            varlen_fwd_config_str = "mha_block_128"
-        elif avg_rows_per_cta > 32:
-            varlen_fwd_config_str = "mha_block_64"
-        elif avg_rows_per_cta > 16:
-            varlen_fwd_config_str = "mha_block_32"
-        else:
-            varlen_fwd_config_str = "mha_block_16"
-        if flag_gems.vendor_name == "mthreads":
-            varlen_fwd_config_str = "mha_block_32"
-
-        cfg = runtime.get_heuristic_config(varlen_fwd_config_str)
-        cfg_params = {
-            "BLOCK_M": cfg["BLOCK_M"](args),
-            "BLOCK_N": cfg["BLOCK_N"](args),
-            "BLOCK_K": triton.next_power_of_2(head_size),
-            # 改动：QK 仍使用完整 BLOCK_K；PV/output 在 D128 split-D 时使用 BLOCK_D=64。
-            "BLOCK_D": 64 if use_varlen_split_d else triton.next_power_of_2(head_size),
-            "SPLIT_D": use_varlen_split_d,
-            "num_warps": cfg["num_warps"](args),
-            "num_stages": 1 if not is_paged else cfg["num_stages"](args),
-        }
-
-        logger.debug("Running flash_varlen_fwd_kernel with config: %s", cfg_params)
-        kernel(*args, **cfg_params)
-
-        if seqlenq_ngroups_swapped:
-            out = out.reshape(
-                batch_size, max_seqlen_q, num_heads_k, head_size
-            ).transpose(1, 2)
-            if out_ is not None:
-                out_.view(batch_size, num_heads_k, max_seqlen_q, head_size).copy_(out)
-                out = out_
-            else:
-                out = out.reshape(batch_size, num_heads_k * max_seqlen_q, head_size)
-            lse = lse.reshape(num_heads_k, batch_size, max_seqlen_q)
-            lse = lse.reshape(num_heads_k * max_seqlen_q, batch_size)
-
-        # unused = torch.empty((), dtype=torch.int64, device=q_device)
-        unused = None
-    return out, q, k, v, lse, philox_args, unused, p
-
-
-def mha_fwd(
-    q,
-    k,
-    v,
-    out,
-    alibi_slopes,
-    p_dropout,
-    softmax_scale,
-    is_causal,
-    window_size_left,
-    window_size_right,
-    softcap,
-    return_softmax,
-    disable_splitkv=False,
-    q_descale=None,
-    k_descale=None,
-    v_descale=None,
-    fp8_p_max=448.0,
-):
-    CHECK_DEVICE(q), CHECK_DEVICE(k), CHECK_DEVICE(v)
-    q_dtype = q.dtype
-    v_dtype = v.dtype
-    q_device = q.device
-    # W8A8 FA3 change: dense forward now follows FA3 FP8 mode: Q/K/V are FP8
-    # tensors and the output keeps the caller's high-precision output dtype.
-    assert q_dtype in _FP8_DTYPES, "W8A8 FA3 FlashAttention expects q to be fp8"
-    assert (
-        k.dtype == q_dtype
-    ), "W8A8 FA3 FlashAttention expects q and k to use the same fp8 dtype"
-    assert v_dtype == q_dtype, "W8A8 FA3 FlashAttention expects v to be fp8"
-    assert q.stride(-1) == 1, "Input tensor must have contiguous last dimension"
-    assert k.stride(-1) == 1, "Input tensor must have contiguous last dimension"
-    assert v.stride(-1) == 1, "Input tensor must have contiguous last dimension"
-    batch_size, seqlen_q, num_heads, head_size = q.size()
-    _, seqlen_k, num_heads_k, _ = k.size()
-
-    # Check output shape
-    if out is not None:
-        assert out.stride(-1) == 1
-        # W8A8 FA3 change: output is explicitly high precision, matching FA3
-        # forward instead of inheriting the FP8 V dtype.
-        assert out.dtype in _HIGH_PRECISION_DTYPES
-        assert out.size() == (batch_size, seqlen_q, num_heads, head_size)
-        CHECK_DEVICE(out)
-
-    assert (
-        head_size % 8 == 0
-    ), "head_size must be a multiple of 8, this is ensured by padding!"
-    assert (
-        num_heads % num_heads_k == 0
-    ), "Number of heads in key/value must divide number of heads in query"
-    if window_size_left >= seqlen_k:
-        window_size_left = -1
-    if window_size_right >= seqlen_k:
-        window_size_right = -1
-    if seqlen_q == 1 and alibi_slopes is None:
-        is_causal = False
-    if is_causal:
-        window_size_right = 0
-
-    is_causal = window_size_left < 0 and window_size_right == 0
-    is_local = window_size_left >= 0 and window_size_right >= 0
-
-    seqlenq_ngroups_swapped = (
-        seqlen_q == 1
-        and alibi_slopes is None
-        and num_heads > num_heads_k
-        and window_size_left < 0
-        and window_size_right < 0
-        and p_dropout == 0
-    )
-    q_groups = num_heads // num_heads_k
-
-    if seqlenq_ngroups_swapped:
-        logger.debug("q_kg swapped.")
-        q = q.reshape(batch_size, num_heads_k, q_groups, head_size).transpose(1, 2)
-        seqlen_q = q_groups
-        num_heads = num_heads_k
-
-    round_multiple = lambda x, m: (x + m - 1) // m * m
-    head_size_rounded = round_multiple(head_size, 32)
-    seqlen_q_rounded = round_multiple(seqlen_q, 128)
-    seqlen_k_rounded = round_multiple(seqlen_k, 32)
-
-    assert (
-        head_size <= 256
-    ), "FlashAttention forward only supports head dimension at most 256"
-    assert head_size == head_size_rounded, "head_size must be rounded to 32"
-
-    def splits_heuristic(num_tasks, num_sms, n_blocks):
-        # splits when wave efficiency is low
-        n_waves = triton.cdiv(num_tasks, num_sms)
-        eff = (num_tasks / num_sms) / n_waves
-        if eff > 0.8 or n_waves > 1:
-            return 1
-
-        min_blocks_per_split = 2
-        best_splits = min(
-            triton.cdiv(n_blocks, min_blocks_per_split),
-            int(math.floor(1.0 / eff)),
-            num_sms,
-        )
-
-        return best_splits
-
-    with torch_device_fn.device(q_device):
-        # Set softmax params
-        lse = torch.empty(
-            (batch_size, num_heads, seqlen_q), dtype=torch.float, device=q_device
-        )
-
-        if out is not None:
-            if seqlenq_ngroups_swapped:
-                out = out.reshape(
-                    batch_size, num_heads_k, q_groups, head_size
-                ).transpose(1, 2)
-        else:
-            # W8A8 FA3 change: allocate a high-precision output even though all
-            # three inputs are FP8.
-            out = torch.empty_like(q, dtype=torch.bfloat16)
-
-        # Set dropout params
-        if p_dropout > 0:
-            is_dropout = True
-            increment = batch_size * num_heads * 32
-            philox_seed, philox_offset = philox_backend_seed_offset(increment)
-            philox_args = torch.tensor(
-                [philox_seed, philox_offset], dtype=torch.int64, device=q_device
-            )
-        else:
-            is_dropout = False
-            philox_args = torch.empty((2,), dtype=torch.int64, device=q_device)
-
-        p_dropout = 1 - p_dropout
-        p_dropout_in_uint8_t = math.floor(p_dropout * 255.0)
-        rp_dropout = 1.0 / p_dropout
-
-        if return_softmax:
-            assert is_dropout, "Only supported with non-zero dropout."
-            p = torch.empty(
-                (batch_size, num_heads, seqlen_q_rounded, seqlen_k_rounded),
-                device=q_device,
-            )
-        else:
-            p = torch.empty((), device=q_device)
-
-        q_nblocks = triton.cdiv(seqlen_q, 128)
-        k_nblocks = triton.cdiv(seqlen_k, 128)
-        # W8A8 FA3 change: per-block descale tensors are part of the kernel
-        # contract.  Benchmarks pass real block scales; default 1.0 keeps manual
-        # FP8 casts runnable while making the scale path visible.
-        q_descale = _normalize_dense_descale(
-            q_descale, batch_size, num_heads, q_nblocks, q_device, "q_descale"
-        )
-        k_descale = _normalize_dense_descale(
-            k_descale, batch_size, num_heads_k, k_nblocks, q_device, "k_descale"
-        )
-        v_descale = _normalize_dense_descale(
-            v_descale, batch_size, num_heads_k, k_nblocks, q_device, "v_descale"
-        )
-
-        M_LOG2E = 1.4426950408889634074
-        if softcap > 0.0:
-            is_softcap = True
-            adjusted_scale_softmax = softcap
-            adjusted_softcap = softmax_scale / softcap
-            adjusted_scale_softmax_log2e = softcap * M_LOG2E
-        else:
-            is_softcap = False
-            adjusted_softcap = 0.0
-            adjusted_scale_softmax = softmax_scale
-            adjusted_scale_softmax_log2e = softmax_scale * M_LOG2E
-
-        # Set alibi params
-        if alibi_slopes is not None:
-            assert alibi_slopes.device == q_device
-            assert alibi_slopes.dtype in (torch.float,)
-            assert alibi_slopes.stride(-1) == 1
-            assert alibi_slopes.shape == (num_heads,) or alibi_slopes.shape == (
-                batch_size,
-                num_heads,
-            )
-            alibi_slopes_batch_stride = (
-                alibi_slopes.stride(0) if alibi_slopes.ndim == 2 else 0
-            )
-            is_alibi = True
-        else:
-            alibi_slopes_batch_stride = 0
-            is_alibi = False
-
-        # ONLY EVEN_K IS SUPPORTED
-        assert head_size == head_size_rounded
-
-        # Do kernel dispatching
-        def dispatch(B, H, Q, K, D, params):
-            num_sms = torch_device_fn.get_device_properties(
-                "cuda"
-            ).multi_processor_count
-            # 改动：D128 使用 split-D dense kernel，把输出维拆成两个 D64 CTA，
-            # 以降低 flash_fwd_kernel 的寄存器压力；splitkv 路径暂不启用 split-D。
-            # use_split_d = D == 128
-            use_split_d = D == 128
-            # 改动：短序列 splitkv 的 combine/临时张量开销通常大于收益，S512 强制走 dense。
-            disable_splitkv1 = disable_splitkv or seqlen_q <= 512
-
-            # Try bh parallel
-            # if B * H > 0.8 * num_sms:
-            #     kernel = flash_fwd_bh_parallel_kernel[(H, B)]
-            #     # Yield kernel and prefilled args
-            #     return kernel, default_args, None, None
-
-            # Try splitkv
-            if (
-                (not use_split_d)
-                and not is_dropout
-                and not is_local
-                and not disable_splitkv1
-            ):
-                BM = block_m_splitkv_heuristic(D)
-                n_tasks = B * H * triton.cdiv(seqlen_q, BM)
-                BN = block_n_splitkv_heuristic(D)
-                n_blocks = triton.cdiv(seqlen_k, BN)
-                n_splits = splits_heuristic(n_tasks, num_sms, n_blocks)
-
-                if n_splits > 1:
-                    logger.debug("kernel: flash_fwd_splitkv")
-                    lse_splits = torch.empty(
-                        (n_splits, B, H, Q), dtype=torch.float, device=q_device
-                    )
-                    out_splits = torch.empty(
-                        (n_splits, B, H, Q, D), dtype=torch.float, device=q_device
-                    )
-                    grid = lambda args: (
-                        triton.cdiv(Q, args["BLOCK_M"]),
-                        n_splits,
-                        B * H,
-                    )
-                    splitkv_kernel = flash_fwd_splitkv_kernel[grid]
-                    params.o_ptr = out_splits
-                    params.softmax_lse_ptr = lse_splits
-                    extra_args = {"blocks_per_split": triton.cdiv(n_blocks, n_splits)}
-                    kernel = splitkv_kernel(*params.args(), **extra_args)
-
-                    if D >= 128:
-                        BLOCK_M = 4
-                    elif D >= 64:
-                        BLOCK_M = 8
-                    else:
-                        BLOCK_M = 16
-                    BLOCK_K = triton.next_power_of_2(D)
-                    grid = lambda args: (triton.cdiv(B * H * Q, BLOCK_M),)
-                    combine_kernel = flash_fwd_splitkv_combine_kernel[grid]
-                    combine_args = {
-                        "out_ptr": out,
-                        "lse_ptr": lse,
-                        "head_size": head_size,
-                        "out_split_stride": out_splits.stride(0),
-                        "lse_split_stride": lse_splits.stride(0),
-                        "out_b_stride": out.stride(0),
-                        "out_s_stride": out.stride(-3),
-                        "out_h_stride": out.stride(-1),
-                        "out_splits_ptr": out_splits,
-                        "lse_splits_ptr": lse_splits,
-                        "n_splits": n_splits,
-                        "BLOCK_M": BLOCK_M,
-                        "BLOCK_K": BLOCK_K,
-                        "q_total": B * H * Q,
-                        "MAX_N_SPLITS": triton.next_power_of_2(n_splits),
-                    }
-                    combine_kernel(**combine_args)
-                    return kernel
-
-            # Last option: flash_fwd
-            logger.debug("kernel: flash_fwd")
-            grid = lambda args: (
-                triton.cdiv(Q, args["BLOCK_M"]),
-                H * B,
-                2 if use_split_d else 1,
-            )
-            kernel = flash_fwd_kernel[grid]
-            # 改动：D128 split-D 时 BLOCK_K 仍为完整 head_dim=128 供 QK 使用，
-            # BLOCK_D=64 只用于 PV 和输出写回；D64/非 split-D 则保持原 BLOCK_K。
-            extra_args = {
-                "BLOCK_D": 64 if use_split_d else triton.next_power_of_2(D),
-                "SPLIT_D": use_split_d,
-            }
-            kernel = kernel(*params.args(), **extra_args)
-            return kernel
-
-        if _debug:
-            p = torch.empty(
-                (batch_size, num_heads, seqlen_q_rounded, seqlen_k_rounded),
-                dtype=torch.float32,
-                device=q_device,
-            )
-            return_softmax = True
-
-        params = fwd_params(
-            q,  # q_ptr,
-            k,  # k_ptr,
-            v,  # v_ptr,
-            out,  # o_ptr,
-            p,  # p_ptr,
-            lse,  # softmax_lse_ptr,
-            q.stride(-3),  # q_row_stride,
-            k.stride(-3),  # k_row_stride,
-            v.stride(-3),  # v_row_stride,
-            q.stride(-2),  # q_head_stride,
-            k.stride(-2),  # k_head_stride,
-            v.stride(-2),  # v_head_stride,
-            out.stride(-3),  # o_row_stride,
-            out.stride(-2),  # o_head_stride,
-            q.stride(0),  # q_batch_stride,
-            k.stride(0),  # k_batch_stride,
-            v.stride(0),  # v_batch_stride,
-            out.stride(0),  # o_batch_stride,
-            False,  # is_cu_seqlens_q,
-            None,  # cu_seqlens_q_ptr,
-            False,  # is_cu_seqlens_k,
-            None,  # cu_seqlens_k_ptr,
-            False,  # is_seqused_k,
-            None,  # seqused_k_ptr,
-            # sizes
-            batch_size,  # b,
-            0,  # bk,
-            num_heads,  # h,
-            num_heads_k,  # hk,
-            num_heads // num_heads_k,  # h_hk_ratio,
-            seqlen_q,  # seqlen_q,
-            seqlen_k,  # seqlen_k,
-            seqlen_q_rounded,  # seqlen_q_rounded,
-            seqlen_k_rounded,  # seqlen_k_rounded,
-            head_size,  # d,
-            head_size_rounded,  # d_rounded,
-            # scaling factors
-            is_softcap,
-            adjusted_softcap,  # softcap,
-            adjusted_scale_softmax,  # scale_softmax,
-            adjusted_scale_softmax_log2e,  # scale_softmax_log2,
-            # W8A8 FA3 change: pass Q/K/V per-block descale metadata to every
-            # compute kernel so QK and PV can both use FP8 tensor cores.
-            q_descale,  # q_descale_ptr,
-            k_descale,  # k_descale_ptr,
-            v_descale,  # v_descale_ptr,
-            q_descale.stride(0),  # q_descale_batch_stride,
-            q_descale.stride(1),  # q_descale_head_stride,
-            q_descale.stride(2),  # q_descale_block_stride,
-            k_descale.stride(0),  # k_descale_batch_stride,
-            k_descale.stride(1),  # k_descale_head_stride,
-            k_descale.stride(2),  # k_descale_block_stride,
-            v_descale.stride(0),  # v_descale_batch_stride,
-            v_descale.stride(1),  # v_descale_head_stride,
-            v_descale.stride(2),  # v_descale_block_stride,
-            True,  # use_fa3_fp8_scales,
-            fp8_p_max,  # fp8_p_max,
-            # dropout
-            is_dropout,
-            p_dropout,
-            rp_dropout,
-            p_dropout_in_uint8_t,
-            philox_args,
-            return_softmax,
-            # causal and swa
-            is_causal,  # is_causal,
-            is_local,  # is_local,
-            window_size_left,  # window_size_left,
-            window_size_right,  # window_size_right,
-            seqlenq_ngroups_swapped,  # seqlenq_ngroups_swapped,
-            False,  # is_paged,
-            # alibi
-            is_alibi,  #
-            alibi_slopes,  # alibi_slopes_ptr,
-            alibi_slopes_batch_stride,  # alibi_slopes_batch_stride,
-            # block table params
-            0,  # total_q,
-            None,  # page_table_ptr,
-            0,  # page_table_batch_stride,
-            0,  # block_size,
-        )
-
-        # Move TxD to last dims for correct stride in Triton tt.load
-        if flag_gems.vendor_name == "iluvatar":
-            params.q_ptr = q.transpose(1, 2)
-            params.k_ptr = k.transpose(1, 2)
-            params.v_ptr = v.transpose(1, 2)
-        kernel = dispatch(batch_size, num_heads, seqlen_q, seqlen_k, head_size, params)
-
-        if _debug:
-            print(f"{kernel.name} shared memory:", kernel.metadata.shared)
-            print(f"{kernel.name} num_warps:", kernel.metadata.num_warps)
-            print(f"{kernel.name} num_stages:", kernel.metadata.num_stages)
-            # print(kernel.asm['ttgir'])
-
-        if seqlenq_ngroups_swapped:
-            out = out.transpose(1, 2).reshape(
-                (batch_size, 1, num_heads_k * seqlen_q, head_size)
-            )
-            q = q.transpose(1, 2).reshape(
-                (batch_size, 1, num_heads_k * seqlen_q, head_size)
-            )
-            lse = lse.reshape((batch_size, num_heads_k * seqlen_q, 1))
-
-        unused = torch.empty((), dtype=torch.int64, device=q_device)
-
-    return out, q, k, v, lse, philox_args, unused, p
